refactor(llm): extract non-native tool-calling into mixin\n\n- New NonNativeToolCallingMixin with pre/post prompt-mock helpers and gating\n- LLM now composes the mixin; calls replaced; old methods removed\n- Keeps behavior: stop-words injection, tool_choice removal, reasoning graft\n- All SDK LLM tests pass under uv (22 tests)\n\nCo-authored-by: OpenHands-GPT-5 <openhands@all-hands.dev>
